refactor: move SecurityHooks into src/hooks
