+++ src/components/ScenarioParameters/components/EditModeButton.js
@@ -8,7 +8,7 @@ import PlayCircleOutlineIcon from '@material-ui/icons/PlayCircleOutline';
 import BackspaceIcon from '@material-ui/icons/Backspace';
 import { useTranslation } from 'react-i18next';
 import { ACL_PERMISSIONS } from '../../../services/config/accessControl';
-import { useUserAppAndCurrentScenarioPermissions } from '../../../state/hooks/SecurityHooks.js';
+import { useUserAppAndCurrentScenarioPermissions } from '../../../hooks/SecurityHooks';
 import { PermissionsGate } from '@cosmotech/ui';
 
 const EditModeButton = ({ classes, handleClickOnDiscardChange, handleClickOnUpdateAndLaunchScenario }) => {
